feat: refactor animation components and introduce new timeline features, including clip management and enhanced controls for improved user experience
